Module Odoc_driver_lib.Run

val instrument : bool

val instrument_dir : Fpath.t lazy_t

type t = {

cmd : string list;

time : float;
(*
Running time in seconds.
*)

output_file : Fpath.t option;

output : string;

errors : string;

status : [ `Exited of int | `Signaled of int ];

}

val commands : t list ref

val n : int Atomic.t

val run : < process_mgr : [> [> `Generic ] Eio.Process.mgr_ty ] Eio.Process.mgr.. > -> Bos.Cmd.t -> Fpath.t option -> t

Return the list of executed commands where the first argument was cmd.

Print an executed command and its time.

val filter_commands : string -> t list

val print_cmd : t -> unit

val k_longest_commands : string -> int -> t list

Returns the k commands that took the most time for a given subcommand.