Module Odoc_driver_lib.Stats

Returns the k commands that took the most time for a given subcommand.

type stats = {

mutable total_units : int Atomic.t;

mutable total_impls : int Atomic.t;

mutable total_mlds : int Atomic.t;

mutable total_assets : int Atomic.t;

mutable total_indexes : int Atomic.t;

mutable non_hidden_units : int Atomic.t;

mutable compiled_units : int Atomic.t;

mutable compiled_impls : int Atomic.t;

mutable compiled_mlds : int Atomic.t;

mutable compiled_assets : int Atomic.t;

mutable linked_units : int Atomic.t;

mutable linked_impls : int Atomic.t;

mutable linked_mlds : int Atomic.t;

mutable generated_indexes : int Atomic.t;

mutable generated_units : int Atomic.t;

mutable processes : int Atomic.t;

mutable process_activity : string Atomic.t Array.t;

mutable finished : bool;

}

val stats : stats

val render_stats : < clock : [> float Eio.Time.clock_ty ] Eio.Time.clock.. > -> generate_json:bool -> int -> unit

val init_nprocs : int -> unit

val pp_stats : Format.formatter -> stats -> unit

val k_longest_commands : string -> int -> Run.t list

val dump : unit -> unit

val compute_min_max_avg : float list -> (float * float * float * int) option

val compute_metric_int : string -> string -> string -> float list -> [> `Assoc of (string * [> `Assoc of (string * [> `Int of int ]) list | `Int of int | `String of string ]) list ] list

val compute_metric_cmd : string -> [> `Assoc of (string * [> `Assoc of (string * [> `Float of float ]) list | `Int of int | `String of string ]) list ] list

val compute_produced_cmd : string -> [> `Assoc of (string * [> `Assoc of (string * [> `Int of int ]) list | `Int of int | `String of string ]) list ] list

Analyze the size of files produced by a command.

val compute_produced_tree : string -> Fpath.t -> [> `Assoc of (string * [> `Assoc of (string * [> `Int of int ]) list | `Int of int | `String of string ]) list ] list

Analyze the size of files outputed to the given directory.

val compute_longest_cmd : string -> [> `Assoc of (string * [> `Assoc of (string * [> `Float of float ]) list | `String of string ]) list ] list

Analyze the running time of the slowest commands.

val all_metrics : Fpath.t -> [> `Assoc of (string * [> `Assoc of (string * [> `Float of float | `Int of int ]) list | `Int of int | `String of string ]) list ] list

val bench_results : Fpath.t -> unit

val total_time : unit -> float